+++ Cosette/Uci/Commands/SetOptionCommand.cs
@@ -60,9 +60,9 @@ namespace Cosette.Uci.Commands
                 { "IIDMinimalDepth", p => SearchConstants.IIDMinimalDepth = int.Parse(p) },
                 { "IIDDepthReduction", p => SearchConstants.IIDDepthReduction = int.Parse(p) },
 
-                { "FutilityPruningMaximalDepth", p => SearchConstants.FutilityPruningMaximalDepth = int.Parse(p) },
-                { "FutilityPruningBaseMargin", p => SearchConstants.FutilityPruningBaseMargin = int.Parse(p) },
-                { "FutilityPruningMarginMultiplier", p => SearchConstants.FutilityPruningMarginMultiplier = int.Parse(p) },
+                { "StaticNullMoveMaximalDepth", p => SearchConstants.StaticNullMoveMaximalDepth = int.Parse(p) },
+                { "StaticNullMoveBaseMargin", p => SearchConstants.StaticNullMoveBaseMargin = int.Parse(p) },
+                { "StaticNullMoveMarginMultiplier", p => SearchConstants.StaticNullMoveMarginMultiplier = int.Parse(p) },
 
                 { "NullWindowMinimalDepth", p => SearchConstants.NullWindowMinimalDepth = int.Parse(p) },
                 { "NullWindowBaseDepthReduction", p => SearchConstants.NullWindowBaseDepthReduction = int.Parse(p) },